requirements.txt
# ¡Salud! Language Learning Platform - Requirements

# Core dependencies
Flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3  # Pinned to be compatible with Flask 2.2.3
pyOpenSSL==23.2.0  # For generating self-signed certificates

# OpenAI API
openai>=1.0.0

# Date and time handling
pytz==2022.7.1

# Database integration
SQLAlchemy==2.0.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
alembic==1.10.2

# User authentication
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
itsdangerous==2.1.2  # For token generation

# API documentation
Flask-RESTful==0.3.9
flasgger==0.9.5

# For testing
pytest==7.2.2
